+++ src/main/java/com/arton/backend/performance/adapter/in/PerformanceController.java
@@ -1,17 +1,9 @@
 package com.arton.backend.performance.adapter.in;
 
-import com.arton.backend.infra.shared.common.CommonResponse;
 import com.arton.backend.infra.shared.common.ResponseData;
-import com.arton.backend.infra.shared.exception.ErrorResponse;
 import com.arton.backend.performance.applicaiton.data.PerformanceInterestDto;
-import com.arton.backend.performance.applicaiton.port.in.PerformanceSearchUseCase;
 import com.arton.backend.performance.applicaiton.port.in.PerformanceUseCase;
 import com.arton.backend.performance.domain.Performance;
-import com.arton.backend.search.adapter.out.persistence.repository.LogRepository;
-import com.arton.backend.search.application.data.SearchPageDto;
-import com.arton.backend.search.application.data.SearchPageDtoV2;
-import com.arton.backend.search.application.data.SearchResultDto;
-import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -20,18 +12,13 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import net.logstash.logback.argument.StructuredArguments;
-import org.jboss.logging.MDC;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Tag(name = "PERFORMANCE", description = "공연 API")
@@ -40,9 +27,6 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PerformanceController {
     private final PerformanceUseCase performanceService;
-    private final PerformanceSearchUseCase performanceSearchService;
-    private final LogRepository logRepository;
-    private final static Logger log = LoggerFactory.getLogger("LOGSTASH");
 
     @Operation(summary = "회원가입시 찜에 필요한 공연 리스트 불러오기", description = "회원가입시 찜에 필요한 공연 리스트를 가져옵니다.")
     @ApiResponses(value = {
@@ -71,78 +55,6 @@ public class PerformanceController {
                 , HttpStatus.OK.value()
                 , performanceService.getOne(id)
         );
-        return ResponseEntity.ok().body(response);
-    }
-
-    @Hidden
-    @PostMapping("/documents")
-    public ResponseEntity<CommonResponse> saveDocuments() {
-        performanceSearchService.saveAllDocuments();
-        return ResponseEntity.ok(CommonResponse.builder().message("ES에 성공적으로 저장하였습니다.").status(HttpStatus.OK.value()).build());
-    }
-
-    @Operation(summary = "공연장소 검색하기", description = "공연장소로 공연을 검색합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "검색 성공",
-                    content = @Content( array = @ArraySchema(schema = @Schema(implementation = SearchResultDto.class))))})
-    @GetMapping("/search/place")
-    public ResponseEntity<ResponseData<Page<SearchResultDto>>> searchByPlace(HttpServletRequest request, @RequestParam(name = "query") String place, @RequestParam(name = "sort", required = false) String sort, @PageableDefault(size = 10) Pageable pageable) {
-        Page<SearchResultDto> documents = performanceSearchService.searchByPlace(place, sort, pageable);
-        MDC.put("keyword", place);
-        log.info("requestURI={}, keyword={}", StructuredArguments.value("requestURI", request.getRequestURI()), StructuredArguments.value("keyword", place));
-        MDC.remove("keyword");
-        ResponseData response = new ResponseData("SUCCESS", HttpStatus.OK.value(), documents);
-        return ResponseEntity.ok().body(response);
-    }
-
-    @Operation(summary = "공연타입 검색하기", description = "공연타입으로 공연을 검색합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "검색 성공",
-                    content = @Content( array = @ArraySchema(schema = @Schema(implementation = SearchResultDto.class))))})
-    @GetMapping("/search/type")
-    public ResponseEntity<ResponseData<Page<SearchResultDto>>> searchByPerformanceType(HttpServletRequest request, @RequestParam(name = "query", required = true) String performanceType, @RequestParam(name = "sort", required = false) String sort, @PageableDefault(size = 10) Pageable pageable) {
-        Page<SearchResultDto> documents = performanceSearchService.searchByPerformanceType(performanceType, sort, pageable);
-        MDC.put("keyword", performanceType);
-        log.info("requestURI={}, keyword={}", StructuredArguments.value("requestURI", request.getRequestURI()), StructuredArguments.value("keyword", performanceType));
-        MDC.remove("keyword");
-        ResponseData response = new ResponseData("SUCCESS", HttpStatus.OK.value(), documents);
-        return ResponseEntity.ok().body(response);
-    }
-
-    @Operation(summary = "공연명 검색하기", description = "공연명으로 공연을 검색합니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "검색 성공",
-                    content = @Content( array = @ArraySchema(schema = @Schema(implementation = SearchResultDto.class))))})
-    @GetMapping("/search/title")
-    public ResponseEntity<ResponseData<Page<SearchResultDto>>> search(HttpServletRequest request, @RequestParam(name = "query", required = true) String query, @RequestParam(name = "sort", required = false) String sort, @PageableDefault(size = 10) Pageable pageable) {
-        Page<SearchResultDto> documents = performanceSearchService.searchByTitle(query, sort, pageable);
-        MDC.put("keyword", query);
-        log.info("requestURI={}, keyword={}", StructuredArguments.value("requestURI", request.getRequestURI()), StructuredArguments.value("keyword", query));
-        MDC.remove("keyword");
-        ResponseData response = new ResponseData("SUCCESS", HttpStatus.OK.value(), documents);
-        return ResponseEntity.ok().body(response);
-    }
-
-    @Operation(summary = "실시간 검색어 화면", description = "실시간 검색어 순위를 보여줍니다.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "검색 성공",
-                    content = @Content(schema = @Schema(implementation = SearchPageDto.class))),
-    @ApiResponse(responseCode = "500", description = "실시간 검색어 집계 실패", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))})
-    @GetMapping("/search")
-    public ResponseEntity<ResponseData<SearchPageDto>> searchLog() {
-        SearchPageDto searchPage = logRepository.getRecentTop10Keywords();
-        ResponseData response = new ResponseData("SUCCESS", HttpStatus.OK.value(), searchPage);
-        return ResponseEntity.ok().body(response);
-    }
-    @Operation(summary = "실시간 검색어 Advanced", description = "실시간 검색어 순위를 보여줍니다. Advanced 버전")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "검색 성공",
-                    content = @Content(schema = @Schema(implementation = SearchPageDto.class))),
-            @ApiResponse(responseCode = "500", description = "실시간 검색어 집계 실패", content = @Content(schema = @Schema(implementation = ErrorResponse.class)))})
-    @GetMapping("/search/v2")
-    public ResponseEntity<ResponseData<SearchPageDtoV2>> searchLogV2() {
-        SearchPageDtoV2 searchPage = logRepository.getAdvancedTop10Keywords();
-        ResponseData response = new ResponseData("SUCCESS", HttpStatus.OK.value(), searchPage);
         return ResponseEntity.ok().body(response);
     }
